requirements.txt
asgiref==3.7.2
cffi==1.15.1
colorama==0.4.6
coverage==7.2.7
cryptography==40.0.2
Django==4.2.1
djangorestframework==3.14.0
iniconfig==2.0.0
jwt==1.3.1
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-env==0.8.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
